Park and Ride Parks in Île-de-France
A Parc Relais is a parking lot set up in the immediate vicinity of a train station or metro station, the purpose of which is to facilitate passenger access to the rail network. In Île-de-France, this equipment mainly concerns the stations furthest from Paris and located in sparse urban environments, where the bus networks cannot meet all the needs of travelers to the stations.

https://data.iledefrance-mobilites.fr/pages/home/